schoolprojs/fall2025/297/BuildFullstackE2EAgentic/personal-expense-assistant/pyproject.toml
[project]
name = "personal-expense-assistant-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.5.0",
    "fastapi>=0.115.12",
    "google-adk>=0.2.0",
    "google-cloud-firestore>=2.20.1",
    "gradio>=5.23.1",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "pydantic-settings[yaml]>=2.8.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
